refactor and cleanup of page.tsx with api error handling
