+++ src/app/page.tsx
@@ -1,97 +1,31 @@
-"use client";
-
-import { useEffect, useState, useMemo } from "react";
 import { Advocate } from "../types/advocate";
+import AdvocateTable from "../components/AdvocateTable";
 
-export default function Home() {
-  const [advocates, setAdvocates] = useState<Advocate[]>([]);
-  const [searchTerm, setSearchTerm] = useState<string>("");
-
-  useEffect(() => {
-    console.log("fetching advocates...");
-    fetch("/api/advocates").then((response) => {
-      response.json().then((jsonResponse) => {
-        setAdvocates(jsonResponse.data);
-      });
-    });
-  }, []);
-
-  const filteredAdvocates = useMemo(() => {
-    if (!searchTerm) return advocates;
+async function getAdvocates(): Promise<Advocate[]> {
+  try {
+    const res = await fetch("http://localhost:3000/api/advocates");
     
-    return advocates.filter((advocate: Advocate) => {
-      const lowerSearchTerm = searchTerm.toLowerCase();
-      return (
-        advocate.firstName.toLowerCase().includes(lowerSearchTerm) ||
-        advocate.lastName.toLowerCase().includes(lowerSearchTerm) ||
-        advocate.city.toLowerCase().includes(lowerSearchTerm) ||
-        advocate.degree.toLowerCase().includes(lowerSearchTerm) ||
-        advocate.specialties.some(specialty => specialty.toLowerCase().includes(lowerSearchTerm)) ||
-        advocate.yearsOfExperience.toString().includes(searchTerm)
-      );
-    });
-  }, [advocates, searchTerm]);
+    if (!res.ok) {
+      throw new Error(`HTTP error! status: ${res.status}`);
+    }
+    
+    const data = await res.json();
+    return data.data || [];
+  } catch (error) {
+    console.error("Error fetching advocates:", error);
+    return [];
+  }
+}
 
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchTerm(e.target.value);
-  };
-
-  const onClick = () => {
-    console.log(advocates);
-    setSearchTerm("");
-  };
+export default async function Home() {
+  const advocates = await getAdvocates();
 
   return (
     <main style={{ margin: "24px" }}>
       <h1>Solace Advocates</h1>
       <br />
       <br />
-      <div>
-        <p>Search</p>
-        <p>
-          Searching for: <span>{searchTerm}</span>
-        </p>
-        <input 
-          style={{ border: "1px solid black" }} 
-          value={searchTerm}
-          onChange={onChange} 
-        />
-        <button onClick={onClick}>Reset Search</button>
-      </div>
-      <br />
-      <br />
-      <table>
-        <thead>
-          <tr>
-          <th>First Name</th>
-          <th>Last Name</th>
-          <th>City</th>
-          <th>Degree</th>
-          <th>Specialties</th>
-          <th>Years of Experience</th>
-          <th>Phone Number</th>
-          </tr>
-        </thead>
-        <tbody>
-          {filteredAdvocates.map((advocate, index) => {
-            return (
-              <tr key ={index}>
-                <td>{advocate.firstName}</td>
-                <td>{advocate.lastName}</td>
-                <td>{advocate.city}</td>
-                <td>{advocate.degree}</td>
-                <td>
-                  {advocate.specialties.map((s, idx) => (
-                    <div key={idx}>{s}</div>
-                  ))}
-                </td>
-                <td>{advocate.yearsOfExperience}</td>
-                <td>{advocate.phoneNumber}</td>
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
+      <AdvocateTable advocates={advocates} />
     </main>
   );
 }
