+++ src/components/AdvocateTable.tsx
@@ -1,0 +1,87 @@
+"use client";
+
+import { useState, useMemo } from "react";
+import { Advocate } from "../types/advocate";
+
+interface AdvocateTableProps {
+  advocates: Advocate[];
+}
+
+export default function AdvocateTable({ advocates }: AdvocateTableProps) {
+  const [searchTerm, setSearchTerm] = useState<string>("");
+
+  const filteredAdvocates = useMemo(() => {
+    if (!searchTerm) return advocates;
+    
+    return advocates.filter((advocate: Advocate) => {
+      const lowerSearchTerm = searchTerm.toLowerCase();
+      return (
+        advocate.firstName.toLowerCase().includes(lowerSearchTerm) ||
+        advocate.lastName.toLowerCase().includes(lowerSearchTerm) ||
+        advocate.city.toLowerCase().includes(lowerSearchTerm) ||
+        advocate.degree.toLowerCase().includes(lowerSearchTerm) ||
+        advocate.specialties.some(specialty => specialty.toLowerCase().includes(lowerSearchTerm)) ||
+        advocate.yearsOfExperience.toString().includes(searchTerm)
+      );
+    });
+  }, [advocates, searchTerm]);
+
+  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchTerm(e.target.value);
+  };
+
+  const onClick = () => {
+    setSearchTerm("");
+  };
+
+  return (
+    <div>
+      <div>
+        <p>Search</p>
+        <p>
+          Searching for: <span>{searchTerm}</span>
+        </p>
+        <input 
+          style={{ border: "1px solid black" }} 
+          value={searchTerm}
+          onChange={onChange} 
+        />
+        <button onClick={onClick}>Reset Search</button>
+      </div>
+      <br />
+      <br />
+      <table>
+        <thead>
+          <tr>
+            <th>First Name</th>
+            <th>Last Name</th>
+            <th>City</th>
+            <th>Degree</th>
+            <th>Specialties</th>
+            <th>Years of Experience</th>
+            <th>Phone Number</th>
+          </tr>
+        </thead>
+        <tbody>
+          {filteredAdvocates.map((advocate, index) => {
+            return (
+              <tr key={index}>
+                <td>{advocate.firstName}</td>
+                <td>{advocate.lastName}</td>
+                <td>{advocate.city}</td>
+                <td>{advocate.degree}</td>
+                <td>
+                  {advocate.specialties.map((s, idx) => (
+                    <div key={idx}>{s}</div>
+                  ))}
+                </td>
+                <td>{advocate.yearsOfExperience}</td>
+                <td>{advocate.phoneNumber}</td>
+              </tr>
+            );
+          })}
+        </tbody>
+      </table>
+    </div>
+  );
+}
